movie_review/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.13.0
djangorestframework-simplejwt>=5.0.0
drf-yasg>=1.20.0
Pillow>=9.0.0
django-filter>=22.1
python-venv
